# Add `PreparedGeometry` to speed up repeated `Relate` operations.