@gauteh 2024-07-08
I am totally unfamiliar with the code, however: From the docs it seems that swap_labels is done depending on whether you do a.relate(b) b.relate(a), which is not known in advance?