@gauteh 2024-07-08
On the other hand, it is weird that this works before this PR? Maybe it is not related.