@michaelkirk 2024-07-08
> From the docs it seems that swap_labels is done depending on whether you do a.relate(b) or b.relate(a), which is not known in advance?

Correct, but `swap_labels` is called on the cloned `GeometryGraph` owned by the context that calls `swap_labels`, so that should be fine. I think the issue is only that GeometryGraph has some internal `Rc`. This was not a relevant before because GeometryGraph was only created as a transient entity. 

The whole point of PreparedGeometry is to keep that entity around for re-use. With this PR it works in a single threaded context. I'd propose that we leave that as the bar for now and follow up with "share a prepared geometry across threads" as future work since this PR is already pretty large.

Plus, that way we can have a focused study of the performance tradeoffs needed (e.g. Rc vs Arc).